requirements.txt
requests
dnspython
colorama
beautifulsoup4
